Only download to first search path directory
Semantically (on UNIX at least) random software shouldn't be writing anywhere
except the user data directories -- its up to the package manager to write to
system ones. As such the metadata available (a fixed path for user dir vs a
search path for system ones) are not conducive to working out which system path
should be used.